requirements.txt
aiogram==3.3.0
aiosqlite==0.19.0
apscheduler==3.10.4
python-dotenv==1.0.0
matplotlib==3.8.2
pytz==2024.1
